+++ src/utils/options.js
@@ -1,0 +1,5 @@
+import parse from './parse.js';
+
+const { config, input, output } = parse(process.argv.slice(2));
+
+export { config, input, output };
